ingestion/producer/requirements.txt
kafka-python>=2.3.0
faker>=24.0.0
